@@ -3,12 +3,14 @@ using SpotifyApp.API.Models;
 
 namespace SpotifyApp.API.Data
 {
-    public class DataContext: DbContext
+    public class DataContext : DbContext
     {
-        public DataContext(DbContextOptions<DataContext> options) : base (options) {}
+        public DataContext(DbContextOptions<DataContext> options) : base(options) { }
 
         public DbSet<Album> Albums { get; set; }
         public DbSet<User> Users { get; set; }
         public DbSet<Photo> Photos { get; set; }
+
+        public DbSet<AlbumRate> AlbumRates { get; set; }
     }
 }
